+++ History.md
@@ -2,5 +2,6 @@
 
 * default to localltunnel.me for host
 * remove exported `connect` method (just export one function that does the same thing)
+* change localtunnel signature to (port, opt, fn)
 
 # 0.2.2 / 2014-01-09